KLIMA-36: Move components out of folders
