+++ assets/react/Components/Logo.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from 'styled-components'
-import Icon from '../Icons/Icons';
+import Icon from './Icons/Icons';
 
 const Logo = () => {
   return (
